refactor: move reducers into separate files
