+++ app/reducers/moviesSkippedReducer.js
@@ -1,0 +1,8 @@
+export default (state = [], action) => {
+  switch (action.type) {
+    case 'SKIP_MOVIE':
+    default:
+      return state;
+  }
+
+}
